fix(config): improve the display of chain names
