+++ package.json
@@ -53,6 +53,7 @@
     "run:bin": "node ./bin/run",
     "run:dev": "node ./bin/dev",
     "clean": "shx rm -rf ./dist",
+    "schemas": "json2ts --input src/repositories/chain.schema.json --output src/types/chain.ts",
     "build": "yarn clean && tsc -b && tsc-alias -p tsconfig.json",
     "lint": "eslint --ext .js,.cjs,.ts --no-error-on-unmatched-pattern ./src",
     "postpack": "shx rm -f oclif.manifest.json",
@@ -128,6 +129,7 @@
     "eslint-plugin-unicorn": "^48.0.1",
     "fancy-test": "^2.0.42",
     "husky": "^8.0.3",
+    "json-schema-to-typescript": "^13.1.1",
     "mocha": "^10.2.0",
     "oclif": "^3.17.2",
     "shx": "^0.3.4",
