+++ scripts/local-1.json
@@ -4,7 +4,7 @@
   "status": "live",
   "network_type": "devnet",
   "website": "https://archway.io/",
-  "pretty_name": "Archway",
+  "pretty_name": "Archway Local Devnet",
   "chain_id": "local-1",
   "bech32_prefix": "archway",
   "daemon_name": "archwayd",
