+++ src/repositories/chain-registry/constantine-3.json
@@ -4,7 +4,7 @@
     "status": "live",
     "network_type": "testnet",
     "website": "https://archway.io/",
-    "pretty_name": "Archway",
+    "pretty_name": "Archway Testnet",
     "chain_id": "constantine-3",
     "bech32_prefix": "archway",
     "daemon_name": "archwayd",
@@ -102,18 +102,30 @@
         {
           "address": "https://rpc.constantine.archway.tech",
           "provider": "Archway Foundation"
+        },
+        {
+          "address": "https://constantine.api.trivium.network:26657",
+          "provider": "TriviumX | Architech"
         }
       ],
       "rest": [
         {
           "address": "https://api.constantine.archway.tech",
           "provider": "Archway Foundation"
+        },
+        {
+          "address": "https://constantine.api.trivium.network:1317",
+          "provider": "TriviumX | Architech"
         }
       ],
       "grpc": [
         {
           "address": "grpc.constantine.archway.tech:443",
           "provider": "Archway Foundation"
+        },
+        {
+          "address": "constantine.api.trivium.network:9090",
+          "provider": "TriviumX | Architech"
         }
       ]
     },
@@ -129,6 +141,12 @@
         "url": "https://testnet.mintscan.io/archway-testnet",
         "tx_page": "https://testnet.mintscan.io/archway-testnet/txs/${txHash}",
         "account_page": "https://testnet.mintscan.io/archway-testnet/account/${accountAddress}"
+      },
+      {
+        "kind": "L0vd",
+        "url": "https://explorers.l0vd.com/archway-testnet",
+        "tx_page": "https://explorers.l0vd.com/archway-testnet/tx/${txHash}",
+        "account_page": "https://explorers.l0vd.com/archway-testnet/account/${accountAddress}"
       }
     ]
   }
