+++ src/types/index.ts
@@ -1,9 +1,9 @@
 export * from './Account';
 export * from './Cargo';
-export * from './Chain';
 export * from './Coin';
 export * from './ConfigData';
 export * from './Contract';
 export * from './ContractTemplate';
 export * from './Deployment';
+export * from './chain';
 export * from './json';
